requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==2.16.0
python-dotenv==1.2.1
pypdf==6.6.2
python-multipart==0.0.12

